feat: added archer tagging
Took 16 minutes

+++ src/main/java/me/lagggpixel/mango/classes/cPlayer.java
@@ -10,7 +10,7 @@ public class cPlayer {
   private Classes classes;
   private double classEnergy;
   private boolean isArcherTagged;
-  private double archerTagTimer;
+  private long archerTagTimer;
 
   public cPlayer(Player player) {
     this.player = player;
